% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/criterion.R
\name{AICc}
\alias{AICc}
\title{Corrected AIC penalty}
\usage{
AICc(ncp, n)
}
\arguments{
\item{ncp}{number of changepoints.}

\item{n}{number of observations.}
}
\description{
Corrected AIC penalty
}
\examples{
AICc(ncp = 2, n = 200)
}
