% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/criterion.R
\name{cp_to_r}
\alias{cp_to_r}
\title{Convert a changepoint index vector to a 0/1 indicator vector}
\usage{
cp_to_r(cp, n)
}
\arguments{
\item{cp}{integer vector of changepoint indices.}

\item{n}{number of observations.}
}
\description{
Convert a changepoint index vector to a 0/1 indicator vector
}
\examples{
cp_to_r(cp = c(5, 12), n = 20)
}
