% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csa.R
\name{csa_theoretical}
\alias{csa_theoretical}
\title{Theoretical CSA under a two-state (stationary/motile) kinetic model}
\usage{
csa_theoretical(theta, s_mesh)
}
\arguments{
\item{theta}{a list with elements `speed_alpha`, `speed_beta`,
`dist_avg`, `p_MM`, `dur_stationary_avg`, `DEPENDENT_SPEED_DUR`.}

\item{s_mesh}{numeric vector of speeds.}
}
\value{
numeric vector, the theoretical CSA evaluated at `s_mesh`.
}
\description{
Theoretical CSA under a two-state (stationary/motile) kinetic model
}
