% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{get_one_plot_size}
\alias{get_one_plot_size}
\title{Compute plot bounds for a single inferred path}
\usage{
get_one_plot_size(path_info)
}
\arguments{
\item{path_info}{a CPLASS output list (with `path_inferred`).}
}
\description{
Compute plot bounds for a single inferred path
}
