% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{get_plot_size}
\alias{get_plot_size}
\title{Compute shared plot bounds across a list of ground-truth paths}
\usage{
get_plot_size(path_list)
}
\arguments{
\item{path_list}{a list of path objects, each with `$path`.}
}
\description{
Compute shared plot bounds across a list of ground-truth paths
}
