% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{get_plot_size_inferred}
\alias{get_plot_size_inferred}
\title{Compute shared plot bounds across a list of inferred paths}
\usage{
get_plot_size_inferred(path_list)
}
\arguments{
\item{path_list}{a list of path objects, each with `$path_inferred`.}
}
\description{
Compute shared plot bounds across a list of inferred paths
}
