% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msd.R
\name{pathwise_MSD}
\alias{pathwise_MSD}
\title{Pathwise mean squared displacement (MSD)}
\usage{
pathwise_MSD(path, use_pct = 0.2)
}
\arguments{
\item{path}{a path object with `path$path$x`, `path$path$y`,
`path$path$t`.}

\item{use_pct}{fraction of the trajectory length to use as the maximum
lag. Default 0.2.}
}
\value{
a tibble with columns `t_lag` and `msd`.
}
\description{
Pathwise mean squared displacement (MSD)
}
