% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_path_inferred_xy}
\alias{plot_path_inferred_xy}
\title{Just the x-y panel of a segmented trajectory}
\usage{
plot_path_inferred_xy(
  path_info,
  xy_width = NA,
  t_lim = NA,
  motor,
  max_speed,
  state_shaded = TRUE
)
}
\arguments{
\item{path_info}{CPLASS output for one path (`path_inferred`,
`segments_inferred`).}

\item{xy_width}{width of the x-y panel; computed automatically if `NA`.}

\item{t_lim}{time-axis limits; computed automatically if `NA`.}

\item{motor}{label used in the plot title.}

\item{max_speed}{upper limit for the speed axis in the duration/speed panel.}

\item{state_shaded}{logical; shade motile/stationary segments. Default `TRUE`.}
}
\description{
Just the x-y panel of a segmented trajectory
}
