% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/criterion.R
\name{r_to_cp}
\alias{r_to_cp}
\title{Convert a 0/1 changepoint indicator vector to changepoint indices}
\usage{
r_to_cp(r)
}
\arguments{
\item{r}{indicator vector.}
}
\description{
Convert a 0/1 changepoint indicator vector to changepoint indices
}
\examples{
r_to_cp(cp_to_r(cp = c(5, 12), n = 20))
}
