% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/criterion.R
\name{sSIC}
\alias{sSIC}
\title{Strengthened Schwarz Information Criterion penalty}
\usage{
sSIC(ncp, n, gamma)
}
\arguments{
\item{ncp}{number of changepoints.}

\item{n}{number of observations.}

\item{gamma}{exponent, \code{gamma > 1} recommended (default in
\code{\link{CPLASS}} is 1.01).}
}
\description{
\eqn{(\log n)^\gamma \rho}, as in Definition 2 of the companion paper.
}
\examples{
sSIC(ncp = 2, n = 200, gamma = 1.01)
}
